# openapi.yaml
...
swagger: "2.0"
info:
  description: Zylitics front end api
  title: Zylitics front end api
  # This version no. is just to showing in the console and has no affect on functionality, although it should be
  # updated whenever production version is updated.
  version: 0.3.9
host: front.api.zylitics.io
...
basePath: /beta
schemes:
  - http # TLS is terminated before hitting ESP, so only http.
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /passwordResets:
    post:
      description: Sends password reset email
      operationId: sendPasswordReset
      parameters:
        - in: body
          name: passwordResetRequest
          required: true
          schema:
            $ref: '#/definitions/PasswordResetRequest'
      responses:
        200:
          description: OK
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /passwordResets/{code}/validate:
    patch:
      description: Validate password reset request
      operationId: validatePasswordReset
      parameters:
        - $ref: '#/parameters/code'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ValidatePasswordResetResponse'
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /passwordResets/{passwordResetId}/resetPassword:
    patch:
      description: Reset password
      operationId: resetPassword
      parameters:
        - in: body
          name: resetPasswordRequest
          required: true
          schema:
            $ref: '#/definitions/ResetPasswordRequest'
        - in: path
          name: passwordResetId
          type: integer
          format: long
          required: true
      responses:
        200:
          description: OK
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /emailChanges:
    post:
      description: Sends email change email
      operationId: sendEmailChange
      parameters:
        - in: body
          name: emailChangeRequest
          required: true
          schema:
            $ref: '#/definitions/EmailChangeRequest'
      responses:
        200:
          description: OK
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /emailChanges/{code}/validate:
    patch:
      description: Validate email change request
      operationId: validateEmailChange
      parameters:
        - $ref: '#/parameters/code'
      responses:
        200:
          description: OK
          schema:
            type: integer
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /emailChanges/{emailChangeId}/changeEmail:
    patch:
      description: Change email
      operationId: changeEmail
      parameters:
        - in: path
          name: emailChangeId
          type: integer
          format: long
          required: true
      responses:
        200:
          description: OK
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
...
  /issue:
    put:
      description: Upload issue file
      operationId: uploadIssueFile
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: file
          type: string # made it string rather than file as endpoint doesn't seem to support file
          required: true
        - in: formData
          name: fileName
          type: string
          required: true
      responses:
        200:
          description: OK
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
    post:
      description: Send issue
      operationId: sendIssue
      parameters:
        - in: body
          name: sendIssueRequest
          required: true
          schema:
            $ref: '#/definitions/SendIssueRequest'
      responses:
        200:
          description: OK
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
...
securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://securetoken.google.com/zl-front-end"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "zl-front-end"
security:
  - firebase: []
parameters:
...
  code:
    in: path
    name: code
    type: string
    required: true
definitions:
...
  PasswordResetRequest:
    type: object
    properties:
      email:
        type: string
  ResetPasswordRequest:
    type: object
    properties:
      password:
        type: string
        format: password
  ValidatePasswordResetResponse:
    type: object
    properties:
      passwordResetId:
        type: integer
        format: int64
      email:
        type: string
...
  EmailChangeRequest:
    type: object
    properties:
      newEmail:
        type: string
...
  ApiError:
    type: object
    required:
      - message
    properties:
      message:
        type: string
...
  SendIssueRequest:
    type: object
    required:
      - desc
    properties:
      desc:
        type: string
      fileName:
        type: string
...